rmp-backend/requirements.txt
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
gunicorn==20.0.4
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.2
mysql-connector-python==8.0.23
newsapi-python==0.2.6
packaging==20.9
pluggy==0.13.1
protobuf==3.15.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.2
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
